# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: LedgerLite API - Financial management for SMEs
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /invoices:
    get:
      operationId: listInvoices
      tags: [invoices]
      summary: List all invoices
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [paid, pending, overdue]
        - name: customerId
          in: query
          schema:
            type: integer
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: List of invoices
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InvoiceList"
    post:
      operationId: createInvoice
      tags: [invoices]
      summary: Create a new invoice
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateInvoiceBody"
      responses:
        "201":
          description: Created invoice
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Invoice"
  /invoices/{id}:
    get:
      operationId: getInvoice
      tags: [invoices]
      summary: Get an invoice by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Invoice details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Invoice"
        "404":
          description: Not found
    put:
      operationId: updateInvoice
      tags: [invoices]
      summary: Update an invoice
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateInvoiceBody"
      responses:
        "200":
          description: Updated invoice
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Invoice"
    delete:
      operationId: deleteInvoice
      tags: [invoices]
      summary: Delete an invoice
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /invoices/{id}/mark-paid:
    post:
      operationId: markInvoicePaid
      tags: [invoices]
      summary: Mark invoice as paid
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Updated invoice
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Invoice"

  # Debts
  /debts:
    get:
      operationId: listDebts
      tags: [debts]
      summary: List all outstanding debts
      parameters:
        - name: customerId
          in: query
          schema:
            type: integer
        - name: aging
          in: query
          schema:
            type: string
            enum: [current, overdue_7, overdue_30, overdue_60]
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: List of debts
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DebtList"
    post:
      operationId: createDebt
      tags: [debts]
      summary: Create a new debt record
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateDebtBody"
      responses:
        "201":
          description: Created debt
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Debt"
  /debts/{id}:
    get:
      operationId: getDebt
      tags: [debts]
      summary: Get a debt record by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Debt details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Debt"
        "404":
          description: Not found
    put:
      operationId: updateDebt
      tags: [debts]
      summary: Update a debt record
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateDebtBody"
      responses:
        "200":
          description: Updated debt
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Debt"
    delete:
      operationId: deleteDebt
      tags: [debts]
      summary: Delete a debt record
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /debts/{id}/mark-paid:
    post:
      operationId: markDebtPaid
      tags: [debts]
      summary: Mark a debt as paid
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Updated debt
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Debt"
...
components:
  schemas:
...
    InvoiceItem:
      type: object
      properties:
        description:
          type: string
        quantity:
          type: number
        unitPrice:
          type: number
        total:
          type: number
      required: [description, quantity, unitPrice, total]
    Invoice:
      type: object
      properties:
        id:
          type: integer
        invoiceNumber:
          type: string
        customerId:
          type: integer
        customerName:
          type: string
        status:
          type: string
          enum: [paid, pending, overdue]
        items:
          type: array
          items:
            $ref: "#/components/schemas/InvoiceItem"
        subtotal:
          type: number
        total:
          type: number
        dueDate:
          type: string
          format: date-time
        paidAt:
          type: string
          format: date-time
          nullable: true
        notes:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, invoiceNumber, customerId, customerName, status, items, subtotal, total, dueDate, createdAt, updatedAt]
    InvoiceList:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Invoice"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
      required: [data, total, page, limit]
    CreateInvoiceBody:
      type: object
      properties:
        customerId:
          type: integer
        items:
          type: array
          items:
            type: object
            properties:
              description:
                type: string
              quantity:
                type: number
              unitPrice:
                type: number
            required: [description, quantity, unitPrice]
        dueDate:
          type: string
          format: date-time
        notes:
          type: string
          nullable: true
      required: [customerId, items, dueDate]
    UpdateInvoiceBody:
      type: object
      properties:
        customerId:
          type: integer
        items:
          type: array
          items:
            type: object
            properties:
              description:
                type: string
              quantity:
                type: number
              unitPrice:
                type: number
            required: [description, quantity, unitPrice]
        dueDate:
          type: string
          format: date-time
        status:
          type: string
          enum: [paid, pending, overdue]
        notes:
          type: string
          nullable: true

    # Debt schemas
    Debt:
      type: object
      properties:
        id:
          type: integer
        customerId:
          type: integer
        customerName:
          type: string
        amount:
          type: number
        amountPaid:
          type: number
        balance:
          type: number
        description:
          type: string
          nullable: true
        dueDate:
          type: string
          format: date-time
          nullable: true
        status:
          type: string
          enum: [active, paid, partial]
        agingDays:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, customerId, customerName, amount, amountPaid, balance, status, agingDays, createdAt, updatedAt]
    DebtList:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Debt"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
      required: [data, total, page, limit]
    CreateDebtBody:
      type: object
      properties:
        customerId:
          type: integer
        amount:
          type: number
        description:
          type: string
          nullable: true
        dueDate:
          type: string
          format: date-time
          nullable: true
      required: [customerId, amount]
    UpdateDebtBody:
      type: object
      properties:
        amount:
          type: number
        amountPaid:
          type: number
        description:
          type: string
          nullable: true
        dueDate:
          type: string
          format: date-time
          nullable: true
        status:
          type: string
          enum: [active, paid, partial]
...