# lib/api-zod/src/generated/types/createInvoiceBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * LedgerLite API - Financial management for SMEs
 * OpenAPI spec version: 0.1.0
 */
import type { CreateInvoiceBodyItemsItem } from "./createInvoiceBodyItemsItem";

export interface CreateInvoiceBody {
  customerId: number;
  items: CreateInvoiceBodyItemsItem[];
  dueDate: Date;
  notes?: string | null;
}