# lib/api-zod/src/generated/types/debtList.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * LedgerLite API - Financial management for SMEs
 * OpenAPI spec version: 0.1.0
 */
import type { Debt } from "./debt";

export interface DebtList {
  data: Debt[];
  total: number;
  page: number;
  limit: number;
}